update Jemkins file

+++ Jenkinsfile
@@ -16,8 +16,13 @@ pipeline {
 		stage('Stop and remove all containers and images') {
 			steps {
 				sh '''cont=$(docker container ps -a | grep -i java | cut -d" " -f1)
-				      docker container stop $cont;docker container rm $cont;docker rmi -f $(docker images -a -q)
-				      echo "All images and container are successfully removed from the Node1"
+					if [ -n "$cont"]
+					then 
+					docker container stop $cont;docker container rm $cont;docker rmi -f $(docker images -a -q)
+				        else
+					docker rmi -f $(docker images -a -q)
+				      	echo "All images and container are successfully removed from the Node1"
+					fi
 				'''
 			}
 		}
